Kinect API update, impl moved to separate package
